split off iohook to own file
